+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Added:
 - Organization and search by manufacturer and model
 - Extended sarch option for own configurations
 - SensorML for devices and platforms now contain a gml:name attribute
+- Campaign field for configurations
 
 Changed:
 - The SMS now sends values for missing information according to the datacite values for unknown information